float vs absolute positioning
Both of them are removed from the normal flow .

position:absolute has no impact  surrounding elements , might overlap them

On the other hand text and inline elements will wrap around the floated element.

The container will not just its height to the element. If child elements are float and no longer on the flow that means the parent element is empty and would be disappered on page